+++ scenes/xiaomi_filter_reset_scene_config.h
@@ -4,7 +4,8 @@
  * @brief Scene list. Included multiple times with different ADD_SCENE definitions.
  */
 ADD_SCENE(xiaomi_filter_reset, start, Start)
-ADD_SCENE(xiaomi_filter_reset, reset, Reset)
+ADD_SCENE(xiaomi_filter_reset, scan, Scan)
 ADD_SCENE(xiaomi_filter_reset, success, Success)
+ADD_SCENE(xiaomi_filter_reset, checked, Checked)
 ADD_SCENE(xiaomi_filter_reset, error, Error)
 ADD_SCENE(xiaomi_filter_reset, about, About)